# Fix infinite loop random sample unique