This is fixed by using `collections.Counter` on the elements (if not already a dict). Duplicate values are then removed so the length of the elements should be the number of unique elements. Using `collections.Counter` counts the occurrence of elements to maintain the higher probability for those values to be in the final sample.

Before this change the following would result in an infinite loop:
```python
from faker import Generator
from faker.providers import BaseProvider
provider = BaseProvider(Generator())
provider.random_sample_unique('aabcd', 5)
```
Because the sample will never be 5 values long because there are only 4 unique values (`'abcd'`) in the provided elements.