@153957 2017-11-02
I also made performance improvements by removing values from the elements which have already added to the sample, because they can not be chosen again.

Before:
```python
%timeit provider.random_sample_unique({'a':1000000, 'e':1}, 2)
# The slowest run took 9.92 times longer than the fastest. This could mean that an intermediate result is being cached.
# 2.72 s ± 1.83 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
After:
```python
%timeit provider.random_sample_unique({'a':1000000, 'e':1}, 2)
# 17.8 µs ± 479 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```